Make c wrapper (#58)
- Added an optional C wrapper
- Added a C version of testapp called "ctestapp"

+++ test/protocol/test_protocol_rx_parsing.cpp
@@ -111,7 +111,7 @@ TEST_F(TestRxParsing, TestRx_Discover_2_parts)
     uint8_t data[12] = {1, 2, 0, 4, 0x7e, 0x18, 0xfc, 0x68};
     add_crc(data, 8);
 
-    for (unsigned int i = 0; i < sizeof(data); i++)
+    for (uint16_t i = 0; i < sizeof(data); i++)
     {
         comm.receive_data(&data[0], i);
         comm.process();
